[SF-0] Added profession stats and main stat calculations

+++ Assets/Scripts/SF/Game/Stats/IStatUpgradeFormula.cs
@@ -2,6 +2,6 @@
 {
     public interface IStatUpgradeFormula
     {
-        float GetStatValue(int baseValue, int tier, int level);
+        int GetStatValue(float baseValue, float tier, int level);
     }
 }
